내보내기 — 오디오 8종 형식과 MIDI
myvocal/audio/export.py

  오디오는 WAV(16/24/32f), FLAC(16/24), MP3, OGG, AIFF 로 내보낸다.
  형식마다 크기가 크게 다르다. 2분 31초 곡이 24비트 WAV 로 42MB, FLAC 16비트로
  14MB, MP3 로 2.2MB 다. 메신저나 업로드 한도에 걸리는 실제 문제라 형식과
  품질을 고를 수 있어야 하고, 고르기 전에 크기를 알려줘야 한다.

  MIDI 는 소리가 아니라 악보로 내보낸다. 다른 DAW 에서 열어 악기를 바꾸거나
  더 편집할 수 있다. 드럼은 관례대로 10번 채널에 놓고, 가사는 lyrics 메타
  이벤트로 넣는다. 읽기도 지원해서 8번 '내 음악 추가' 의 입력으로 쓸 수 있다.

고친 버그
  긴 곡을 OGG 로 내보내면 프로세스가 통째로 죽었다. 2분 31초(740만 샘플)를
  한 번에 인코더에 넘기면 세그멘테이션 폴트가 났다. 예외가 아니라 프로세스가
  죽는 것이라 try/except 로도 못 잡는다. 10초 단위로 나눠 쓰도록 바꿔서
  해결했고, 덤으로 메모리 사용량이 곡 길이와 무관하게 일정해졌다.
  8개 형식 전부 155초 곡에서 크래시 없이 통과한다.

  MIDI 메타 이벤트가 latin-1 로 인코딩돼서, 한글 트랙 이름이나 가사가 있으면
  저장이 통째로 실패했다. UTF-8 로 쓰고, 읽을 때는 UTF-8 을 먼저 시도한 뒤
  안 되면 latin-1 로 되돌아가게 했다.

  choose_bpm 이 round() 로 int 를 돌려줘서 BPM 타입이 섞였다. MIDI 는 템포를
  정수 마이크로초로 저장하므로 왕복하면 소수가 되는데, int 와 비교하면
  항상 어긋난다.

검증
  tests/test_export.py  114개 항목.
  손실 없는 형식은 파형이 그대로인지 샘플 단위로 확인하고, 긴 곡 내보내기는
  자식 프로세스에서 돌려 죽더라도 검증이 멈추지 않게 했다. MIDI 는 2976개
  음이 음높이·위치·길이·세기까지 하나도 어긋나지 않고 돌아오는지 확인한다.

  크기 예상 검증에는 잡음과 타악기 성분을 섞은 신호를 쓴다. 순수 사인파만
  쓰면 FLAC 이 90% 넘게 줄여버려서 압축 형식의 예상값을 검증할 수 없다.

+++ myvocal/music/composer.py
@@ -118,7 +118,10 @@ def choose_bpm(profile: GenreProfile, generator: random.Random) -> float:
     low, high, typical = profile.bpm_low, profile.bpm_high, profile.bpm_typical
     # 보통값을 중심으로 한 삼각 분포
     value = generator.triangular(low, high, typical)
-    return round(value)
+    # 정수로 반올림하되 float 로 돌려준다. BPM 은 소수가 될 수 있는 값이고
+    # (MIDI 는 템포를 정수 마이크로초로 저장하므로 왕복하면 소수가 된다),
+    # 여기서 int 를 돌려주면 타입이 섞여서 비교할 때마다 걸린다.
+    return float(round(value))
 
 
 def select_instruments(
